YAID: Yet Another Icon Display
There are so many icons which can be used in labels and properties
but there is no real good way to know what that icon looks like
without actually using it (none that I figured out anyway). That's
where YAID becomes useful. It's a small add-on which displays all
available icons.

Couple of issues with the addon. I don't know how to get the icon
enumeration from the code so I can iterate through it. What I did
is copy the icon list from the manual and am iterating over that.

Ideally, we should be able to get the enumeration in the code and
iterate over that. This way if the enumeration changes in the future
we can simply pick it up.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>